:package: Allow changing video and loop times. Add show/hide button.
Implements a function to take the values from the form inputs
and actually do something with them: update the player object.
Users are now able to choose their own video by video ID, and
they can also modify the start/end times for the portion of
the video they want to loop.

This commit also introduces a new button to show or hide the
video player. Its onclick function checks the 'display' CSS
value and sets it to either none or block. The audio of the
video will keep playing in the background. If you are on
mobile, this also works and you're still able to play/pause
via the notification on Android (presumably something equivalent
on iOS would also work). Note that this does break seeking
on the Android notification, which I'll need to look into.

The show/hide video feature is excellent when you pair it with
locking your phone and just leave the audio playing. I have
not benchmarked it yet, but hopefully it'll help save battery!