:wrench: Add a bunch of frontend libraries to the project
Adds ZURB Foundation For Sites, jQuery and what-input to the project.
These will be needed for UI work and to make things pretty and
generally usable and decent. jQuery and what-input are dependencies
for Foundation, but we would probably need jQuery somewhere down
the line anyway.

Updates the index.html to make use of these frameworks. This
made the page look weird, but don't worry, it'll be phixed with
some borrowed code from our very own phixyn.com website, which
coincidentally uses Foundation too! No way...