:package: Add a two-handle slider for loop time input
Adds a Foundation two-handle slider that can be used to set
the start and end times of the portion of the video to loop.
This can be used in addition to the number input fields. The
input fields will update a new number of seconds everytime
the slider is moved, as the slider handles are bound to the
input fields. Similarly, if a new value is entered in the
input fields, the slider is updated.

This is very _epico_. Just needs some bug fixing and
adjustments.