:package: Implement a simple HTTP server to act as a proxy
Implement a very simple Python HTTP server to perform certain requests on the
behalf of our app. This is needed to bypass CORS: an annoying, and kinda
useless IMO, mechanism that prevents a web application running at one origin,
access to selected resources from a different origin. If we write another
application (that is not a web application) to perform the request and pass
it back to our web application, we can get around this mechanism.

In this case, the application performing the request is Python with
urllib.request. The data received in the response is sent to our web
application using a Python HTTP server. Both of these Python components are
currently in this one Python application that I refer here as 'proxy'.

In the near future, I will probably use websockets for IPC between the Python
application and the web application, eliminating the use of http.server. This
is both due to security reasons (http.server shoult NOT be used in production,
pls don't ever do this) and speed reasons. Also, websockets are modern and
shiny.