:package: Use Python application to get around CORS and improve UX
See parent commit for more detailed explanation. TL;DR is that calling
the update slider and inputs function when the player changed state
to PLAYING would often cause the end slider/time value to reset back
to the last second of the video. This happened if the user paused
the video and then unpaused it - we would look at the player's
getDuration() again and set endTime to that. This commit fixes this.

Non TL;DR explanation:

A much nicer UX would require set endTime to be set only once, when
the video is first queued (when the user clicks the submit button
for new video ID). However, nothing in the API allowed us to do this,
as we could only get a valid video duration AFTER the video starts
playing.

As a workaround, we could make the same API request that YouTube
makes when the play button is pressed for the first time. If we
made this request before the video is played, we could get the
video info and duration. However, due to the CORS mechanism, we
couldn't do this from the web application.

To do this and provide a nice UX for our users, we needed to proxy
the request to another application (NOT a web application) and then
have that application send us back the video length. The parent
commit implements that application, and this commit adds code to
request the data from that application and make use of it, thus
finally giving us a nice UX and fixing the issue with the end
time changing when it shouldn't. This is probably the biggest and
longest bug fix I've ever done in my life tbfh.

Finally, this commit also includes some refactors and improvements,
mainly around the function to update the slider and inputs, but
that's less relevant than the essay above.