:bug: Fix player not seeking to startTime if current time was lower
Fixes an issue wherein, if the user seeked the video to a time before
the startTime value, the player would just keep playing from that
point and would not go back to the original startTime. For now, I
would prefer it if the player always stays within the set loop
portion, and if the user wants to seek around, then they should
set start/end to the beginning and end of the video.