:package: Add buttons to set start/end times to current video time
Updates the webpage with two new buttons. These can be used to set
the start/end times to the video's current time. It is a really
useful way to find precise start and end times to create the perfect
loop, without getting frustrated at the sliders or having to try to
guess the exact number of seconds you need to put in the input
fields.

Pro tip: While the player is paused and focused, press . and , to
navigate between frames in the video. When I switch to using floats
for the start/end times, it will be possible to set the **exact**
frames to start and end a loop on :fire: