Make input fields and buttons darker (WIP)
Begin improving dark theme form element colors. This makes the
input fields and the primary buttons darker. Still not 100%
happy with this color pallete. Also, there's a weird box-shadow
color because I had to override Foundation's with something. It'll
definitely need to be changed.

The :focus color needs to be changed next, but I think the buttons
look decent. The input background also looks kinda decent, but not
100% on it. Also, main text color should change to #afafaf in the
future.