Update :focus colors for text input boxes
Make the text input boxes have a more subtle :focus background color
and lighter border color. Instead of relying on the background color
to make the element "pop", rely on the border. It feels a bit easier
on the eyes and smoother? Also the text inside the box might be
easier to read, not sure.