Update custom property names in CSS
Update the names of the custom properties used for colors in the
CSS file. This is to make the naming more consistent and also to
better describe the use of the properties. Some colors have also
been re-used where it made sense, and others removed.

This ensures that we're limiting the number of colors we're using
to a sensible number, instead of having a million different shades
of grey. It makes the style more consistent, and theming easier.
With these changes, there are at most 3 different colors for each
purpose (primary, secondary and accent).