Remove collapsing responsive menu
Remove the menu bar that would appear in place of the top bar when
on mobile. Since we have no links in the menu, it did not make
sense to have this. This also removes part of the FOUC workaround
from the CSS, leaving only what's needed for the top bar.

As part of this, the font size of the site title was also made
responsive, as otherwise it would look ugly on mobile. It now
uses the vw (viewport width) unit to make it responsive. Other
bits of its styling were updated to make it look nicer on mobile
and desktop.

This closes #25.