Try out using theme accent as :focus border color
Instead of having an accent color for borders that is light grey,
use the theme's accent color (currently hot pink) for this instead.
Also, make the width of input:focus borders 2px instead of 1px, as
this seems common in websites like Twitch and YouTube. I'm not sure
if I like it yet and it seems to only look good on mobile, but
maybe I'll get used to it.