Fix footer not sticking to the bottom of the page
Fixes an issue wherein the footer would appear in the middle of the
page if the page didn't have enough content to fill it up vertically.
The issue was fixed by adding a "main-content" ID to the first
grid-container div, which is technically the main content wrapper.
Then, its min-height was set to 100vh (viewport height, which is the
height of the browser window). Now this div will always take up the
height of the window, therefore pushing any further elements (such
as the footer) down the page even if the div's children don't take
up the whole height.

While looking up documentation to help solve this issue, I realized
that the footer doesn't need its own grid-container, and that having
it in a grid-container could cause more issues. Therefore, it has
been removed as part of this fix. It's legal to have grid-x and
grid-y outside grid-container. In fact, grid-container is an
optional class used to constrain the width of the grid and its
elements. It didn't make sense to use it for the footer, as we were
just using the default grid-x/y behaviour by applying the 'full'
class anyway.

Lastly, as part of this, the old CSS "hack" has been removed, as
this fix is a much more modern and efficient solution. The footer
was given a margin-top of 5em.

This closes #8.