Implement state storage and updating in frontend client
Update app.js to store a state object. This contains the videoId,
startTime and endTime. The state object gets updated a certain
amount of time after the start/end time(s) are updated (still need
to make it update on new video ID).

The state is stored in the browser via the History API, with
replaceState(). This function also allows adding querystring to
the URL, so that has been implemented. See code comments for
more info on implementation details and rationale.