Move some logic into .ready() and window "load" event handlers
Move a bunch of initialization logic to jQuery's ready() and window
"load" event handlers. jQuery's ready is fired when the DOM is ready
to be manipulated. The "load" event is fired when all elements on a
page have finished loading, including iframes.

This fixes an issue wherein the "player" variable would be accessed
before it was defined. It also fixes an issue wherein some callback
functions would be called twice or not at the right time. Generally,
this is much cleaner and might even improve performance a bit, as
well as being good practice.