Call updateState() after setting state object in .ready()
Make a call to updateState() after setting the state object in the
jQuery .ready() handler. This might not always be needed, but I saw
some issues without this call. Mainly, there could be a chance that
history.state never gets set, and this causes issues with the URL
not updating + state object not being correct.