Move registering of slider event handlers to player ready handler
So there seems to be a bit of a race condition on when jQuery's
"ready" and window "load" events are fired. Their firing has been
made asynchronous (was previous synchronous) and it makes no sense.
There's a GitHub issue for it, the devs don't seem to give a toss
and they consider it an anti-pattern to do the suggested workarounds
for when you want to work with these events synchronously.

See: https://github.com/jquery/jquery/issues/3194

Now, since we can't rely on "load" being fired after "ready" like
it should fucking be in the first place, we gotta move some logic
that was in "load" to somewhere else. I wanted the logic inside
"load" to fire whenever the DOM was FULLY LOADED (iframe and all),
but jQuery devs think that is an edge case or some shit, idk man.
Luckily, there's still a way to fire things when the iframe is
loaded, and that is by putting code inside the YouTube player's
"ready" handler. That's right mofos, jQuery stupidity ain't stopping
us.

So all the logic that is supposed to execute when the DOM is FULLY
LOADED (aka what the "load" event should be) has been moved to
player ready. Can we guarantee that the player ready event will
always be fired when the whole DOM is done loading? Not sure, but
it's the best we're gonna get right now, so suck it up I guess and
blame jQuery devs.