Add a workaround for slider fill bug
Adds a workaround for an issue wherein the slider fill will be
the wrong width after the video's end time changes from its
initial value. This bug would occur just after the page was fully
loaded, or after the player finished loading a new video. There's
currently two known workarounds for the issue: call jQuery's
.change() on either of the input elements, or call Foundation's
_reflow() on the loop portion slider. The latter seems a bit cleaner,
but it's important to note that it will also cause a change event
to be fired.

This is a known issue with the Foundation slider plugin, as
documented here:
https://github.com/foundation/foundation-sites/issues/10768

The drawback of our workaround (apart from being ugly with a
setTimeout) is that when the change event is fired, we call
updateState() (because it's in the slider's changed event handler).
This has no harm, but it's unnecessary code being called twice, and
it may actually make the page slower for a few milliseconds - not
noticeable, but undesirable anyway. Also, it just feels like we
shouldn't have to do this at all and the slider plugin should sort
its shit out (it would be good if the Foundation devs fixed it tbh).

For now, I'll probably leave the workaround and make a GitHub issue
for this.