Client: Fix full videos sometimes not looping
Fixes an issue wherein looping full videos would sometimes not work.
This happened everytime a new video was loaded by the user. The
reason for the bug was that we loaded a single video whenever we
updated the player with the user's new video ID. In order for full
videos to loop, we need to load a playlist with the video ID and
then call setLoop(true). This was the behaviour already for the
first video loaded when the page first loaded, because of the
'playlist' and 'loop' attributes being set in 'playerVars'.

For a full explanation see the code comment in this commit's diff.

Closes #4.