Client: Reduce slider reflow delay from 2 to 1 second
With the slider fill bug workaround, reduce the time we wait to
call reflow() from 2 seconds to 1 second. Having tested this for
quite a bit, I'm confident that it can be done without negative
effects. We might even be able to go lower, just have to play
around with it.