Client: Allow users to enter a YouTube video URL to load a new video
Refactor updatePlayer() to handle YouTube video URLs in addition to IDs. Makes
use of the previously implemented functions (see parent commit) to check if
the string in the input field is a URL.

For video IDs, an additional check has been put in place to somewhat validate
the ID (just by checking its length for now).

If the URL does not contain a video ID, or if the video ID is invalid, then the
function returns to exit, and nothing is sent to the WebSocket server. However,
errors are only logged, not shown to the user yet.