Form: Use Foundation Abide plugin to validate inputs in form
Integrate Abide in the HTML form to validate the input fields. Errors will now
be displayed to users if they enter and try to use invalid data, such as
letters in the numeric fields. Errors may also be shown for invalid video
links, but this still needs to be implemented on app.js. Help messages have
also been added beneath each field.

Abide is a form validation library that extends the HTML5 validation API with
custom validators.

Docs:
https://get.foundation/sites/docs/abide.html