Core: Validate start and end time before updating state
Start and end time values are now validated to ensure they are a valid number.
The state object is only updated with valid values, and not NaN. The Abide
plugin API is used to add HTML classes to invalid input fields and to show
a relevant error message beneath these fields.

With this change, start and end times in the state object are only updated
when their value is different (e.g. startTime != state.start).