Client: Move numeric input validation to slider moved event listener
It made more sense to put logic related to slider movement in the slider moved
event listener. Note that this is mainly used to clear any error classes from
the numeric inputs once the slider is moved, since the slider always sets
numbers (or at least it should o.O).