Form: Remove data-form-error-on attribute from start and end inputs
While all we need on a numeric field is the "required" error, this assumes
that every browser will always unset/remove any value that's not a number.
Even though Firefox and Chrome do this, it's better to err on the side of
caution and catch all errors. Thus, only catching "required" errors has
been removed, and we should now be catching both pattern (non-number) and
required errors.