Client: Update websocket onmessage handler to handle new server payloads
Updates the client to work the the server changes introduced in the parent
commit. Adds a check for a "type" property in the message. If the check
passes, a switch statement is used to look for and use the video info in the
message. If no supported content is found in the message, an error is logged
to the console, along with the message received.