Client: Add an event handler for the focus event on input elements
Adds an event listener and handler function for the "focus" event of HTML
`<input>` elements. The handler function checks if the element has a custom
data attribute called "autoselect" (data-autoselect). If it does and it's set
to true, then the text value of the input will be automatically selected when
it gains focus.

This uses the jQuery object's data() function to parse the value of the
data-autoselect attribute. It automatically converts the string value to a
boolean, or undefined if the element doesn't include the attribute.

Note that `$(this)` is only available through the function()'s context and
will not be available with an Arrow function. As it turns out, the same
applies to the document ready event handler - this will be addressed in the
next commit.