Looper: Remove "Copy" button from share link input group
The copy button does not work on mobile unless the website has a valid SSL
certificate. To avoid having functionality that only works on desktop, it
has been removed for the time being until a better solution to handle this
can be found.

This commit also makes the share link auto-select when the input field gains
focus, to make it easier to copy.