Client: Update input focus event handler to handle autocopy data attribute
Update the event handler for `<input>` focus events to handle a new data
attribute called "data-autocopy". This attribute specifies whether the text
value of the input field should be automatically copied to the user's
clipboard.

If `autocopy` is found to be `true`, the Clipboard API is used to attempt to
copy the input's value to the user's clipboard. This will fail on mobile if
the site does not use a valid SSL certificate, for security purposes.
However, if `autoselect` is used, the users can still copy the text with one
tap, as the selected text's context menu will appear with a "Copy" button.

This uses the same logic that was in the "copyShareLink()" function. This
function has now been removed, along with its unused variables.