Looper: Automatically copy share link to clipboard when users click on it
When users click on the input field containing the share link, it is
automatically copied to the user's clipboard. This works on desktop as long
as the current tab is the active tab. On mobile, this will only work if the
site uses a valid SSL certificate and the user has allowed their browser app
access to their clipboard. However, on mobile, a context menu will be shown
to the user when they tap on the input field, as its contents are
automatically selected, and they will be able to just tap "Copy".

This is really the best solution so far for a cross-platform/mobile-friendly
implementation of automatically copying the link to clipboard.