Looper: Remove fade-in animation when opening share link modal
The fade-in and out animations are too slow and trying to configure them
like the documentation says is simply not working. Tried to look for
potential fixes for the animation speed, but none of them work and the
motion-ui docs and Foundation docs are not helpful. Unfortunately, the
animation has to be removed as it's simply garbage. For now, the slow
fade-out animation has been left in, but if it proves to be too frustrating,
it will also be removed.

On a side note, always be updating your documentation please.