Looper: Display 'link copied' in help text for 2.5s after share link is copied
As we currently don't have a tooltip component, we need a different way to
tell desktop users about the super secret exclusive auto-copy feature. This
commit adds a hack that looks for the help text element (usually containing an
'aria-describedby' attribute) and temporarily changes its text to "Link
copied!". This lasts for 2 and a half seconds before the help text is set back
to its original.