Styles: Add styling for 'lights off' overlay
Style the 'Turn Off the Lights' overlay div. This is the first version of it,
needs some more testing. As part of this, z-index values were updated to make
sure there are no clashes with the Foundation reveal component (see details
below).

Other misc changes:

- Move player container to custom components section
- Update all z-index values to use lower values
- Make sure player is still dimmed when a Foundation reveal is shown (done
by updating the z-index values as stated above)