Looper: Fix player container element showing while video is hidden
Since we introduced a shadow to the player container element, hidding the
player element would result in the shadow remaining visible.

This commit changes the togglePlayer() function to toggle the player
container instead of the player element. For this to work properly, it
has to change the opacity, rather than using fadeToggle(). This is
because fadeToggle() actually changes the display property, causing
the space taken by the container to vanish. If we toggle the visible or
opacity properties instead, the space taken by the container remains,
and the feature behaves exactly as it did before the shadow was added.