Fix end time input and slider being stuck at 1 second
Fixes an issue wherein the value of the "End time" input field could get
stuck at 1 second along with the slider.

The issue was due to a race condition caused by a call to the YouTube IFrame
API. A call to `player.getDuration()` would often return 0. This value would
then be used to set the input fields' max value attributes, and the slider's
logical and visual end values.

According to the YouTube IFrame API documentation, this can happen if
`getDuration()` is called before the video metadata finishes loading. While we
have been able to reliably use `getDuration()` for months in the
`onPlayerReady` event handler, this stopped being reliable this week. It is
not clear why, as there have been no documented changes in the API.

This commit removes the call to `getDuration()`, replacing it with sending a
message to our backend server asking for the video duration. This is the same
approach used in the `updatePlayer()` function, called when the user wants to
load a new video, since we had the same problem there.

- - -

About `isInitialVideo`:

Since the `onmessage` handler for our websocket would also set the `state.end`
to the video duration, some changes had to be made to prevent side-effects.
Importantly, we don't want the `state.end` value to change from its original
value if the user is visiting using a URL they had bookmarked for a specific
loop. Instead, we want to use the end value in the querystring of their URL.
As such, a slight change has been made in the `onmessage` handler and
`updatePlayer()` - see the comment about `isInitialVideo` in the diff.