Refactor: Remove unnecessary and commented out code
Remove the call to endTimeInput.val() in the websocket's onmessage handler, as
it has no effect whatsoever and is not needed as a workaround for issue #54.
Remove the commented out code for issue #54 in the onPlayerReady handler, as
well as the workaround for issue #52 which has been moved to the websocket
onmessage handler and had no effect being called in the onPlayerReady handler.