Fix 'Update' button not working when used for the first time
Fixes an issue wherein the video player would not load a new video after the
'Update' button was clicked or tapped. This would only happen the first time
that users would click/tap the 'Update' button.

This commit changes the way the first playlist is cued when the page is
loaded. Instead of using 'playlist' in 'playerVars' when constructing the
player object, we use the player.cuePlaylist() API function in the
`onPlayerReady` event handler.

This bug happened because the first playlist was not being cued. Calling
cuePlaylist or loadPlaylist once would "force" the playlist to cue, and
subsequent playlists would load without any issue. Previously, we have
been able to rely on the 'playlist' property being set in 'playerVars' to
cue the first playlist. However, this seems to no longer be the case, and
cuePlaylist() has to be used instead.

The reason we need to cue and load playlists instead of single videos is to
ensure that videos loop forever. By combining cue/loadPlaylist with setLoop,
videos will always loop after reaching the end. A single video could stop
playing after 1 loop, especially on mobile with the phone locked. For more
context, see PR #57.