Update outline color for buttons on mobile
Changes the outline color for buttons in the :active, :focus and :hover to a
light gray color that matches the theme. This change is aimed at mobile
browsers, particularly Chrome and Safari, that set the outline color to yellow
in their User-Agent stylesheets.

While the original outline color is ugly, it is necessary for accessibility.
Instead of removing it completely, it is better to change the color to one
that matches the color scheme. This commit also changes the background color
of tapped buttons on mobile (usually blue) to a light gray transparent color.