YT API: Update get_video_info endpoint (old hotfix)
This commit adds an old hotfix applied on my local instance, which I never
committed to the git repo (oops). The get_video_info endpoint started
responding with a non-OK status (status 410 I think) to all requests.

When I went to a video on YouTube and looked at the outgoing requests in
dev tools, I saw that the endpoint was still being used but with different
query string parameters. I added those parameters to the request being made
and it all worked fine for a few weeks. Eventually, this endpoint stopped
working completely and yet another solution was found (see next commit).