YT API: Hotfix for getting video duration
I think these are the same changes already in a separate branch,
so I will copy the commit message:

The infamous get_video_info endpoint is no longer working, so we need an
alternative way to get video duration. We can actually get this from the
video player object. However, the way this hotfix/workaround is implemented
in this commit is not great/quite slow.

For now, the alternative HTTP request will be implemented in a separate
branch, but this branch and commit should be kept in the repo and re-visited
later. Maybe it is a way to get rid of the nonsense of needing a server for
this project.

Will prioritize merging this branch rather than the old one.