Implement a custom dual-handle range slider
Create a no-BS slider for the Looper. This is part of the effort
to get rid of Foundation as a dependency, reduce bloat, and also
reduce BS!

However, it is also necessary for the removal of jQuery (#139), as
the current form inputs and slider are tied together with Foundation,
which is tied together to jQuery.

This commit only creates the slider, it doesn't refactor the existing
code. The slider was tested on a separate dev-only page. The full
refactor will happen as part of #137 and #139.

Part of #137.